fix: move snapshot state function to base_service
